winter_crypto

Module hashers

Contains implementations of currently supported hash functions.

Structs

Blake3_192

Implementation of the Hasher trait for BLAKE3 hash function with 192-bit output.

Blake3_256

Implementation of the Hasher trait for BLAKE3 hash function with 256-bit output.

Rp62_248

Implementation of Hasher trait for Rescue Prime hash function with 248-bit output.

Rp64_256

Implementation of Hasher trait for Rescue Prime hash function with 256-bit output.

RpJive64_256

Implementation of Hasher trait for Rescue Prime hash function with 256-bit output.

Sha3_256

Implementation of the Hasher trait for SHA3 hash function with 256-bit output.